requirements.txt
aiogram==3.24.0
aiosqlite==0.22.1
pydantic-settings==2.12.0
SQLAlchemy==2.0.46
alembic==1.18.1
numpy==2.4.1
redis==7.1.0
pytest==9.0.2
google-auth==2.48.0
gspread==6.2.1